[depthMap] Remove MultiViewParams and ImagesCache dependencies in DeviceCamera
